+++ src/features/item-list/components/index.ts
@@ -1,0 +1,1 @@
+export * from './item-list-section';
