+++ src/features/item-list/index.ts
@@ -1,3 +1,4 @@
 export * from './apis';
+export * from './components';
 export * from './model';
 export * from './ui';
